+++ images/base-notebook/start-singleuser.py
@@ -20,5 +20,5 @@ if "NOTEBOOK_ARGS" in os.environ:
 command += sys.argv[1:]
 
 # Execute the command!
-print("Executing: " + " ".join(command))
+print("Executing: " + " ".join(command), flush=True)
 os.execvp(command[0], command)