+++ tests/by_image/base-notebook/test_healthcheck.py
@@ -5,10 +5,34 @@ import time
 
 import pytest  # type: ignore
 
-from tests.utils.get_container_health import get_health
 from tests.utils.tracked_container import TrackedContainer
 
 LOGGER = logging.getLogger(__name__)
+
+
+def get_healthy_status(
+    container: TrackedContainer,
+    env: list[str] | None,
+    cmd: list[str] | None,
+    user: str | None,
+) -> str:
+    container.run_detached(
+        environment=env,
+        command=cmd,
+        user=user,
+    )
+
+    # giving some time to let the server start
+    finish_time = time.time() + 10
+    sleep_time = 1
+    while time.time() < finish_time:
+        time.sleep(sleep_time)
+
+        status = container.get_health()
+        if status == "healthy":
+            return status
+
+    return status
 
 
 @pytest.mark.parametrize(
@@ -60,22 +84,7 @@ def test_healthy(
     cmd: list[str] | None,
     user: str | None,
 ) -> None:
-    running_container = container.run_detached(
-        tty=True,
-        environment=env,
-        command=cmd,
-        user=user,
-    )
-
-    # giving some time to let the server start
-    finish_time = time.time() + 10
-    sleep_time = 0.1
-    while time.time() < finish_time:
-        time.sleep(sleep_time)
-        if get_health(running_container) == "healthy":
-            return
-
-    assert get_health(running_container) == "healthy"
+    assert get_healthy_status(container, env, cmd, user) == "healthy"
 
 
 @pytest.mark.parametrize(
@@ -108,22 +117,7 @@ def test_healthy_with_proxy(
     cmd: list[str] | None,
     user: str | None,
 ) -> None:
-    running_container = container.run_detached(
-        tty=True,
-        environment=env,
-        command=cmd,
-        user=user,
-    )
-
-    # giving some time to let the server start
-    finish_time = time.time() + 10
-    sleep_time = 0.1
-    while time.time() < finish_time:
-        time.sleep(sleep_time)
-        if get_health(running_container) == "healthy":
-            return
-
-    assert get_health(running_container) == "healthy"
+    assert get_healthy_status(container, env, cmd, user) == "healthy"
 
 
 @pytest.mark.parametrize(
@@ -145,18 +139,6 @@ def test_not_healthy(
     env: list[str] | None,
     cmd: list[str] | None,
 ) -> None:
-    running_container = container.run_detached(
-        tty=True,
-        environment=env,
-        command=cmd,
-    )
-
-    # giving some time to let the server start
-    finish_time = time.time() + 5
-    sleep_time = 0.1
-    while time.time() < finish_time:
-        time.sleep(sleep_time)
-        if get_health(running_container) == "healthy":
-            raise RuntimeError("Container should not be healthy for this testcase")
-
-    assert get_health(running_container) != "healthy"
+    assert (
+        get_healthy_status(container, env, cmd, user=None) != "healthy"
+    ), "Container should not be healthy for this testcase"