+++ tests/by_image/base-notebook/test_start_container.py
@@ -42,14 +42,10 @@ def test_start_notebook(
         f"Test that the start-notebook.py launches the {expected_command} server from the env {env} ..."
     )
     host_port = find_free_port()
-    running_container = container.run_detached(
-        tty=True,
-        environment=env,
-        ports={"8888/tcp": host_port},
-    )
+    container.run_detached(environment=env, ports={"8888/tcp": host_port})
     # sleeping some time to let the server start
     time.sleep(2)
-    logs = running_container.logs().decode()
+    logs = container.get_logs()
     LOGGER.debug(logs)
     # checking that the expected command is launched
     assert (
@@ -76,10 +72,9 @@ def test_tini_entrypoint(
     https://superuser.com/questions/632979/if-i-know-the-pid-number-of-a-process-how-can-i-get-its-name
     """
     LOGGER.info(f"Test that {command} is launched as PID {pid} ...")
-    running_container = container.run_detached(tty=True)
+    container.run_detached()
     # Select the PID 1 and get the corresponding command
-    cmd = running_container.exec_run(f"ps -p {pid} -o comm=")
-    output = cmd.output.decode().strip("\n")
+    output = container.exec_cmd(f"ps -p {pid} -o comm=")
     assert "ERROR" not in output
     assert "WARNING" not in output
     assert output == command, f"{command} shall be launched as pid {pid}, got {output}"