+++ tests/by_image/docker-stacks-foundation/test_python_version.py
@@ -14,7 +14,6 @@ def test_python_version(container: TrackedContainer) -> None:
     )
     logs = container.run_and_wait(
         timeout=5,
-        tty=True,
         command=["python", "--version"],
     )
     python = next(line for line in logs.splitlines() if line.startswith("Python "))
@@ -28,7 +27,6 @@ def test_python_pinned_version(container: TrackedContainer) -> None:
     LOGGER.info(f"Checking that pinned python version is {EXPECTED_PYTHON_VERSION}.*")
     logs = container.run_and_wait(
         timeout=5,
-        tty=True,
         command=["cat", "/opt/conda/conda-meta/pinned"],
     )
     assert f"python {EXPECTED_PYTHON_VERSION}.*" in logs