+++ tests/utils/conda_package_helper.py
@@ -29,7 +29,6 @@ from collections import defaultdict
 from itertools import chain
 from typing import Any
 
-from docker.models.containers import Container
 from tabulate import tabulate
 
 from tests.utils.tracked_container import TrackedContainer
@@ -41,57 +40,31 @@ class CondaPackageHelper:
     """Conda package helper permitting to get information about packages"""
 
     def __init__(self, container: TrackedContainer):
-        self.running_container: Container = CondaPackageHelper.start_container(
-            container
-        )
+        self.container = container
+        self.container.run_detached(command=["bash", "-c", "sleep infinity"])
+
         self.requested: dict[str, set[str]] | None = None
         self.installed: dict[str, set[str]] | None = None
         self.available: dict[str, set[str]] | None = None
         self.comparison: list[dict[str, str]] = []
 
-    @staticmethod
-    def start_container(container: TrackedContainer) -> Container:
-        """Start the TrackedContainer and return an instance of a running container"""
-        LOGGER.info(f"Starting container {container.image_name} ...")
-        return container.run_detached(
-            tty=True,
-            command=["bash", "-c", "sleep infinity"],
-        )
-
-    @staticmethod
-    def _conda_export_command(from_history: bool) -> list[str]:
-        """Return the mamba export command with or without history"""
-        cmd = ["mamba", "env", "export", "--no-build", "--json"]
-        if from_history:
-            cmd.append("--from-history")
-        return cmd
-
     def installed_packages(self) -> dict[str, set[str]]:
         """Return the installed packages"""
         if self.installed is None:
             LOGGER.info("Grabbing the list of installed packages ...")
-            self.installed = CondaPackageHelper._parse_package_versions(
-                self._execute_command(
-                    CondaPackageHelper._conda_export_command(from_history=False)
-                )
-            )
+            env_export = self.container.exec_cmd("mamba env export --no-build --json")
+            self.installed = CondaPackageHelper._parse_package_versions(env_export)
         return self.installed
 
     def requested_packages(self) -> dict[str, set[str]]:
         """Return the requested package (i.e. `mamba install <package>`)"""
         if self.requested is None:
             LOGGER.info("Grabbing the list of manually requested packages ...")
-            self.requested = CondaPackageHelper._parse_package_versions(
-                self._execute_command(
-                    CondaPackageHelper._conda_export_command(from_history=True)
-                )
+            env_export = self.container.exec_cmd(
+                "mamba env export --no-build --json --from-history"
             )
+            self.requested = CondaPackageHelper._parse_package_versions(env_export)
         return self.requested
-
-    def _execute_command(self, command: list[str]) -> str:
-        """Execute a command on a running container"""
-        rc = self.running_container.exec_run(command, stderr=False)
-        return rc.output.decode()  # type: ignore
 
     @staticmethod
     def _parse_package_versions(env_export: str) -> dict[str, set[str]]:
@@ -101,7 +74,7 @@ class CondaPackageHelper:
         # since we only manage packages installed through mamba here
         # They are represented by a dict with a key 'pip'
         dependencies = filter(lambda x: isinstance(x, str), dependencies)
-        packages_dict: dict[str, set[str]] = dict()
+        packages_dict: dict[str, set[str]] = {}
         for split in map(lambda x: re.split("=?=", x), dependencies):
             # default values
             package = split[0]
@@ -126,7 +99,7 @@ class CondaPackageHelper:
             )
             # Keeping command line output since `mamba search --outdated --json` is way too long ...
             self.available = CondaPackageHelper._extract_available(
-                self._execute_command(["mamba", "search", "--outdated", "--quiet"])
+                self.container.exec_cmd("conda search --outdated --quiet")
             )
         return self.available